+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+### Changed
+- Rename the "Crop to largest area" option to "Crop to union of detected objects". Previously saved crop settings still load.
+- [Dev] Extract crop planning from the main window into a `crop_planner` module.
+
 ## [0.9.1] - 2026-06-25
 ### Changed
 - [Dev] Make release script handle HTTPS git remotes.